// frontend/src/api/openapi/models/RefineTopicModelParams.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RefineTopicModelParams = {
  /**
   * Type of the TMJob
   */
  tm_job_type?: string;
  /**
   * ID of the aspect to refine.
   */
  aspect_id: number;
};